refactor: replaced old types for activity pool
